+++ Source/Frontend/UI/UI.csproj
@@ -424,6 +424,7 @@
     <Compile Include="UIConnector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Extensions\ComponentForm.cs" />
+    <Compile Include="Extensions\DataGridViewDraggable.cs" />
     <Compile Include="Extensions\DataGridViewNumericUpDownCell.cs" />
     <Compile Include="Extensions\DataGridViewNumericUpDownColumn.cs" />
     <Compile Include="Extensions\DataGridViewNumericUpDownEditingControl.cs" />
